+++ _randomgen/doc/source/change-log.rst
@@ -4,6 +4,8 @@ Change Log
 Changes since v1.14
 ===================
 
+- Add legacy generator which allows NumPy replication
+- Improve type handling of integers
 - Switch to array-fillers for 0 parameter distribution to improve performance
 - Small changes to build on manylinux
 - Build wheels using multibuild